tirpc: Fix numerous errors and warnings.
Apart from minor printf() formatting fixes and removal of unused
variables this commit adds "#include <pthread.h>" to testcases that are
utilizing pthreads and fixes argument passing to the pthread start
function.

There should be no changes in functionality except for one testcase that
was missing "return test_status;" at the end of the main().

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>