syscalls/read02: Fix failures on Btrfs
Btrfs fallbacks to buffered I/O when file is opened with O_DIRECT and
bufffer or length is not aligned and therefore such reads works fine.

To quote Jan Kara:

Historically what O_DIRECT means is: "Make my Oracle DB run faster."
which was not well defined for the cases which the DB did not trigger
and different people then came to different solutions for the
cornercases.

So this patch changes the O_DIRECT testcases to pass either if read()
failed with EINVAL or if the buffer was read() fine.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>