tst_kvercmp: Do not segfault on short release number
Debian testing (jessie) has "3.14-2-amd' in the release string and
because of that all testcases that call tst_kvercmp() segfaults in
atoi(NULL).

So if there is no third number in the version string, we now assume
(possibly incorrectly) that it's zero. But at least these testcases runs
fine there.

There seems to be three number version as a part of the uname.version
string there but parsing that seems to be even more fragile.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>