syscalls/getgroups/getgroups01: set supplementary groups to prevent TCONF
On a system where root belongs to only one group, the 3# test of
getgroups01 will fail by TCONF. But if we call setgroups() to set
supplementary groups, the test can succeed if run as privileged user.

With this patch, it'll still fail by TCONF when being run by
unprivileged user.

Signed-off-by: Han Pingtian <hanpt@linux.vnet.ibm.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>
Acked-by: Jan Stancek <jstancek@redhat.com>