openposix: Fix fork-11-1 hangs
The testcase was hanging on some systems.

The cause was locked stdout which caused deadlock at buffer flush at the
exit() in child process and because of that the parent ended up waiting
in wait() forever.

This commit cleans up the source a bit, gets rid of the testfrmw and
adds funlockfile at the end of the threaded function which fixes the
deadlock.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>