runltp: export initialized LTP_DEV
commit 5c7c3fb219957484db92111d423e4e2553fb580b removed initialization
of empty LTP_DEV from setup(), but it also caused that LTP_DEV
is no longer exported when it's initialized in set_block_device().

This causes every test to attempt to get loop device on its own,
which is not necessary and may presumably sporadically cause issues
on older buggy kernels during cleanup of loop device.

This patch makes LTP_DEV exported (when it's not empty) and
tst_acquire_device() will use that instead of trying to acquire/release
loop device on its own for each test.

Signed-off-by: Jan Stancek <jstancek@redhat.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>