lib/tst_fs_link_count.c: check fs type and do not fill subdir if no limit
This number of max sub-directories depends on the filesystem. For current
kernel, subdir limit is not availiable for all filesystems (availiable for
ext2, ext3, minix, sysv etc).

So we check the fs type in tst_fs_fill_subdirs(), and only try to reach the
limit on filesystems which have subdir limit. Otherwise, give an information
tells that subdir limit is not avaliable, and returns 0 directly.

Signed-off-by: Sheng Yong <shengyong1@huawei.com>