sendmsg01: remove ".flags = ~MSG_CMSG_COMPAT" case
Testing sendmsg syscall with multiple flags can lead to undefined results,
depending on order in which these flags are processed.

Futher, this flag is defined by LTP in different way than it's defined
in kernel, which causes testcase to fail as it hits different check
in kernel for 64 and 32 bit testcase.

This patch removes the case entirely. And since this is the only place
using it, it removes msg_common.h, where it's defined as well.

Signed-off-by: Jan Stancek <jstancek@redhat.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>