sched_setparam/10-1: remove testcase
As in 9-1, this hangs on -rt kernel, because alarm(2)
never fires.

In case it runs on single cpu system, it's pretty
much same as 9-1: child process of higher priority
preempts parent.

In case it runs on 2+ CPUs, it doesn't matter that
one child is able to preempt parent, shared value
is going to be increased regardless in all other
children.

Signed-off-by: Jan Stancek <jstancek@redhat.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>