kernel modules: fix syntax errors when modules are disabled
If there's no kernel sources provided, we end up with build time warnings:
$ make
/bin/sh: line 0: [: too many arguments
/bin/sh: line 0: [: too many arguments
...

These don't break the build as the end result is the same -- the modules
get skipped.  But we can easily avoid these warnings by checking the vars
in the makefiles before we finally do the version compare in the shell.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>