fsstress: Kill children only if test was interrupted.
This fixes the -l option that has been broken since the backport from
xfstest:

commit 0841eee5c88f42edc02794591e0bd6fb55596145
Author: Zorro Lang <zlang@redhat.com>
Date:   Tue Sep 2 23:45:58 2014 +0800

    fs/fsstress: backport fix from xfstests

The problem is that fsstress in xfstest does not support the -l option at all
and sends the SIGTERM to the whole process group unconditionally, which caused
parent to receive the signal as well and the shoud_stop flag was set from
signal handler which makes the xfstest exit immediately.

This patch fixes #32.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>