syscalls: select04, poll02: Use pipe instead of stdin
Using stdin for the timeouting select()/poll() was bad idea after all.
Not only it exits the test when executed from command line when any key
is pressed before it finishes, but it also fails with the SUSE test
execution framework.

This patch changes these to make use of read end of the pipe instead.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>