aiocp: Add block size detection.
The testcase fails with O_DIRECT on drives with block sizes > 512
(which was hardcoded in the test source).

This commit adds detection for the drive block size and forces the right
aligment and file size if needed.

It also aborts the test if requested block size is not multiple of drive
block size. (The test exits with 0 in this case, TCONF would be better
but we would have to fix the rest of the test to be compatible with the
LTP library first.)

The failures were found on s390x where DASD has 4k block size.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>