lib: tst_device: Do not TBROK on EACCES
Do not call tst_resm(TBROK, ...) if we fail to open "/dev/loop-control"
with EACCES because of lacking priviledges.

The thing is that there are testcases that want to try to fallback to
loop device if /tmp is not suitable for the testing. And while they do
not require root for the testing, attaching loop devices is priviledget
operation.

We do not want to use tst_require_root() in these for several reasons.

The first reason is that we cannot know if we would need to used loop
devices before tst_tmpdir() is called, hence we would need to do cleanup
before the test exits. And while tst_require_root() has cleanup
parameter, it's deprecated.

The second reason is that we do not want to skip rest of the testcases
just because we cannot use loop devices. The testcase was simply
skipping one part of the testing previously.

In the end changing this case from TBROK to TINFO is easiest solution.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>