controllers/memcg: Fix limit alignment expectations
since 3e32cb2e0a12 ("mm: memcontrol: lockless page counters") kernel
commit the limit_in_bytes is round down rather than up. This behavior
change has been discussed during the review and it was considered a
reasonable so fix the test case. As we have to be backward compatible
we have to check for both round down and up. It doesn't make any sense
to give the target value in all callers so calculate the appropriate
values in test_limit_in_bytes.

Signed-off-by: Michal Hocko <mhocko@suse.cz>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>