controllers: Fix --mmap-lock1 test
After kernel 3.12 doing a mmap with the MAP_LOCKED flag
wouldn't trigger an oom. This has been discussed and explained
in http://marc.info/?l=linux-mm&m=142122902313315&w=2

To work around this simply touch the memory after
mmap, that way we are going to cause a fault and the process
is going to be killed. This fixes both use_hierarchy and
the memcg functional tests

Signed-off-by: Nikolay Borisov <n.borisov@siteground.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>