tst_rmdir: do msync(MS_SYNC) before munmap()
This is a follow up to commit 57027655ce2d021ccbe7db7172847cafafa30d60
("Unmap the futexes backend file in tst_rmdir").

As D8 in the NFS FAQ (http://nfs.sourceforge.net) states, in order
to be 100% sure that all modified pages were written to the file system
we need to employ msync(MS_SYNC).

Signed-off-by: Stanislav Kholmanskikh <stanislav.kholmanskikh@oracle.com>