containers: new testcase userns03
ID-outside-ns is interpreted according to which process is opening
the file. If the process opening the file is in the same user namespace
as the process PID, then ID-outside-ns is defined with respect to the
parent user namespace. If the process opening the file is in a different
user namespace, then ID-outside-ns is defined with respect to the user
namespace of the process opening the file.

Signed-off-by: Yuan Sun <sunyuan3@huawei.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>