syscalls/mprotect04: Fix for ia64
IA64 ABI calls functions by function descriptors and because of that we
cannot simply copy function address since the address we get in C is
addres of the function descriptor and not the function itself.

Instead of that we copy a code for a minimal function to the mmaped page
and create a function descriptor for it.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>