syscalls/include/: delete linux_syscall_numbers.h
Delete linux_syscall_numbers.h in LTP otherwise it will not be
regenerated from git checkout even if the header is not up to date.

The reason why it was added to git was that the *.in files does not
change too frequently and that the script that generated them took up to
minute of time. But since the script was fixed in:

commit a0dd3db40468c9a34c39fc9ea03eeb1ca280e0cd
Author: Jan Stancek <jstancek@redhat.com>
Date:   Tue Sep 8 09:58:11 2015 +0200

    linux_syscall_numbers.h: speed up generation

There is no reason to keep the header in tree anymore.

Signed-off-by: Cui Bixuan <cuibixuan@huawei.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>