Fix including config.mk and features.mk in env_pre.mk
If you try to run make in the clean sources tree then it leads
to the infinite loop.

So check if config.mk or features.mk exists before inclusing and
don't use "-include" form. The problem is that inclusion of these
files may lead to the infinite loop if these files don't exist.
It happens because config.mk and features.mk have recipes and
Make tries to generate them and include them in env_pre.mk again
and again until resources are exhausted.

Signed-off-by: Yuriy Kolerov <yuriy.kolerov@synopsys.com>
Reviewed-by: Jan Stancek <jstancek@redhat.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>