containers/share: added namespaces types argument for ns_exec
In contrast with all other namespaces, joining the same user namespace
returns EINVAL. Because of that ns_exec now requires to specify the list
of types of namespaces in which to execute a program. The list of a comma
separated namespaces should be in the same order as list specified for
ns_create.

All testcases in containers/netns were modified to use namespaces types
argument when executing ns_create and ns_exec.

netns/netns_sysfs.sh:
* renamed "dummy0" to less common name "dummy_test0"
* /sys is now correctly mounted as private in a new namespace before
  testing to not affect main mount namespace
* added test case which checks that only localhost and dummy interfaces
  are seen in /sys/class/net/ after mounting sysfs in a new namespace

Signed-off-by: Matus Marhefka <mmarhefk@redhat.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>