diotest6: Cleanup & fix.
* Small cleanup

* Fixed heap corruption (as reported by Sandy Sun)

The code incorrectly passed iovec pointers instead of buffers (which
worked most of the time, since usually the pointer to structure is the
same as pointer to first member). But this may and will fail. Moreover
the test doesn't need the iovec array at all, two simple buffers do
suffice.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reported-by: Sandy Sun <a22017@motorola.com>