include/affinity.h: define _GNU_SOURCE on Linux
Otherwise the affinity cannot be used.

This is better than setting _GNU_SOURCE in the testcases
unconditionally, the affinity.h header must be include first though.

Another option would be compiling the affinity function into object file
and linking it with testcases (which wouldn't need to expose the GNU
extensions in the testcases).

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>