mtest01: fail if a child unexpectedly terminates
mtest01 checks that a specified amount of memory could be allocated.
The memory allocation is performed by several child processes. Each
child allocates certain amount of memory and then sends SIGRTMIN to
the parent. The parent waits until it receives SIGRTMIN from all child
processes or until the amount of free memory is decreased by at least
the amount it tasked the children to allocate.

The problem is that in certain situations the child process gets killed
by oom-killer. The parent never receives SIGRTMIN from that process and
there is not sufficient decrease in the amount of free memory, so
the parent keeps waiting indefinitely.

This patch fixes the above issue. It modifies the test in such a way that
it also waits for SIGCHLD signal (which is sent when the oom-killer
terminates the child process). If SIGCHLD is received the test FAILS.

Signed-off-by: Jiri Vohanka <jvohanka@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>