set_thread_area01: fix build issue
get_thread_area() and set_thread_area() system calls are not supported
on a set of targets. Corresponding syscall numbers are not defined, and
in LTP they become defined as '-1' in <linux_syscall_numbers.h> header.

This test uses switch-case for its own purposes with _NR_get_thread_area
and __NR_set_thread_area as case selections. It causes compile-time
error, at least for ARM target.

In this patch, the test code is refactored to remove switch-case, and
repair test build.

Warning about comparsion of signed and unsigned fixed too.

Signed-off-by: Yury Norov <ynorov@caviumnetworks.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>