module.mk: Skip module build properly
As it was the foo.ko module may end up in MAKE_TARGETS (for any make
goal) even if WITH_MODULES was set to 'no' if there was any file with
.ko suffix left in the directory. Since the SKIP was set to 2
incorrectly in that case.

Not that this was a serious bug but it caused quite a lot of head
scratching before I figured out why the build system is even trying to
build the modules when kernel-devel package wasn't installed and the
configure set WITH_MODULES to 'no'.

Now we explictilty add the '*.ko' into MAKE_TARGETS only for install and
clean goals instead of playing with different SKIP variable values.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>