build: regen.sh (linux_syscall_numbers.h): Fix race
The file is set as global make dependency in testcase.mk and the problem
is that the regen.sh script still takes 0.2s on reasonably modern
machine and because of that the parallel make can run the script several
times since the file is created at the very end of the script. I've seen
build failures where the script was stared from both tools/Makefile and
testcases/kernel/Makefile and the build failed since the result was
corrupted header.

So this patch adds '@touch linux_syscall_numbers.h' before we start the
regen.sh script so that the rest of make processes will see the goal as
finished and wouldn't start the regen.sh script while one instance is
allready running.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>