kernel/input: Redo open_device()
The opendir approach si prone to race conditions. Since if the newly
created input char device was missing at the time we called opendir() we
will never try to open it.

This changes the function to open /dev/input/event[0-100] blindly
instead.

We also retry opening the device with a short usleep() in between to
make sure we do not fail in cases where the device creation is delayed
substantially.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>