commands/mkswap01: Fix race
It can take up to 10ms for the /dev/disk/by-* files to appear after we
do the mkswap -L or mkswap -U. Now the test waits up to 100ms for the
files to be created before it calls swapon.

This fixes random sporadic failures I've seen on some systems.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>