test.sh: Add tst_umount(), use it in mkfs and df tests
For the same reason we added it to the C library in 420add2.

To quote the commit to the C library:

"Various desktop deamons (gvfsd-trash is known for that) may be
 stupid enough to probe all newly mounted filesystem which results in
 umount(2) failing with EBUSY."

The shell version does also check if the device is actually mounted and
do not attempt to do umount if it's not so it could be directly used in
cleanups.

This commit also makes use of the tst_umount() in mkfs01.sh and df01.sh
which has been found failing rarely.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>