syscalls/renameat202: improve renameat2_verify()
In renameat2_verify(), str is allocated on stack with size 8 (sizeof
content) and filled with random data. Then file content (7 bytes) is
read into str and the last byte in str is left unchanged with garbage,
and test fails if that last byte is not zero.

Another problem is, as Cyril Hrubis pointed out, the str buffer is too
small, so there's chance to overrun the str buffer, as read(2) is
reading strlen(content) + 10 bytes.

Fix them by extending str buffer size and initializing str with zeros.
Also improved the checking logic, do strncmp only if expected number of
bytes is read from file. And improved the failure messages too.

Suggested-by: Cyril Hrubis <chrubis@suse.cz>
Signed-off-by: Eryu Guan <eguan@redhat.com>