commands/mkswap: Do not TWARN on missing symlink
Apparently the symlink is not the only way how to find a device by its
label. Some versions of swapon use libblkid to probe first 512 bytes of
each block device as well. Unfortunately this is not always case and
there are systems that will fail when the symlink wasn't created yet and
there are systems that does not create the symlink at all.

So the best we can do here is to change the TWARN to TINFO so that the
test works fine in a case that symlink was not created but swapon probes
block devices to get the right device for a given label.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Signed-off-by: Stanislav Kholmanskikh <stanislav.kholmanskikh@oracle.com>
Signed-off-by: Shuang Qiu <shuang.qiu@oracle.com>