mprotect04: invalidate icache on powerpc
Kshitij Malik reported that this testcase is crashing on his
PowerPc based board MPC8360 with SIGILL on PROT_EXEC test
when it tried to execute copy of exec_func.

Coherency and Synchronization Requirements for PowerQUICC III, 2.2 says:
"Instruction cache coherency must be handled separately from data cache
coherency. Whereas there is hardware support for data cache coherency,
instruction cache coherency must be maintained in software. Even if
stores are performed to pages marked as coherence required, icbi
instructions are required to ensure the instruction cache is coherent."

And indeed, we were able to confirm, that invalidating icache resolved
the problem. Without patch, it was highly reproducible, with patch it
ran fine for minutes in a loop. Presumably mmap() mapped an area that
was previously a code, that was still in icache.

The patch is a powerpc assembly, but it should be safe, because we
already do this in different testcase since:
commit 038e4bc450d1 ("POWER5 has coherent icache, but POWER4,
PPC970 and some other processors lack it.")

This patch has some extra syncs, I updated it according to current
kernel version from:
  https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/tree/arch/powerpc/lib/code-patching.c#n19

Reported-and-tested-by: Kshitij.Malik@mitel.com
Signed-off-by: Jan Stancek <jstancek@redhat.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>