include/lapi/fcntl.h: Do not include tee and splice
The newly added fcntl34 test uses newlib and includes lapi/fcntl.h. The
fcntl header includes fallback defitions for splice and tee which are
exposed on older distros. The fallback defitions calls ltp_syscall()
that calls tst_brkm() which is not available under the newlibrary and
hence the compilation fails.

This commit removes the tee.h, splice.h, and vmsplice.h from fcntl.h and
adds these directly to the testcases if needed.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>