syscall/fcntl33.c: fix test on FUSE fs
Even though a child process is waiting for a parent to be in 'S'
state before triggering SIGIO, the signal might come earlier than
sigtimedwait() being invoked in the parent process. As a result the
parent waited for the event that had already happened.

I can only think about that the parent process may be in a sleep state
somehow while executing open() syscall on which it might sleep while
running it on FUSE FS:
vfs -> fuse kernel driver -> userspace implementation.

It can be fixed if we move open() before fork.

Signed-off-by: Alexey Kodanev <alexey.kodanev@oracle.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>